+++ CHANGELOG.md
@@ -31,6 +31,11 @@ all APIs might be changed.
   selected on a composite.
 - Cynic will now error if a QueryFragment selects no fields.
 
+### Changes
+
+- The `graphql_type` parameter for most of the derives is now optional.  It 
+  defaults to the name of the struct/enum if not present.
+
 ## v0.12.2 - 2020-02-22
 
 ### Bug Fixes